Move Venue and Court entities into VenueAggregate
