+++ src/TournamentManagement/TournamentManagement.Domain/TournamentAggregate/Tournament.cs
@@ -3,6 +3,7 @@ using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Linq;
+using TournamentManagement.Domain.VenueAggregate;
 
 namespace TournamentManagement.Domain.TournamentAggregate
 {
