+++ src/TournamentManagement/TournamentManagement.Domain/VenueAggregate/Venue.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Linq;
 
-namespace TournamentManagement.Domain
+namespace TournamentManagement.Domain.VenueAggregate
 {
 	public class Venue : Entity<VenueId>
 	{
